backend/requirements-backend.txt
# SuraSmart Backend - Phase 1 Requirements
# Core Framework
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
django-timezone-field==6.1.0

# Caching & Message Queue
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Authentication & Authorization
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1

# API & Serialization
django-extensions==3.2.3

# ML/AI (Phase 1 & Future)
deepface==0.0.93
opencv-python==4.8.1.78
numpy==2.0.2
tensorflow==2.14.0  # For future ML models
torch==2.1.1  # Alternative ML framework

# Image Processing
Pillow==10.1.0

# Server & Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Monitoring & Logging
sentry-sdk==1.39.1

# Testing (Development)
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality (Development)
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Documentation (Development)
drf-spectacular==0.27.0

# Development
django-debug-toolbar==4.2.0
